Support for anonymous functions
If the first element of an expression is not a keyword then it is
evaluated, rather than calling lookup() explicitly. If it is a function
bound to a variable then eval_in_env() will perform lookup() and return
the closure, otherwise the lambda will be evaluated and return the
closure directly.